GetInventoryItem(i)

+++ Fallout1/FalloutLib/include/Fallout1/RE/InventoryItem.h
@@ -4,7 +4,7 @@
 
 namespace Fallout1::RE {
     struct InventoryItem {
-        uint32_t ptr;
-        uint32_t quantity;
+        uintptr_t ptr;
+        uint32_t  quantity;
     };
 }
